# Conflicts:
#	src/main/java/com/am/adastra/entity/vo/UserHistorySimpleVO.java
#	src/main/resources/mapper/UserHistoryMapper.xml

+++ src/main/java/com/am/adastra/entity/dto/VideoOperateDTO.java
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
-import java.util.Date;
 
 /**
  * @Author : ArturiaMu KMUST-Stu
@@ -21,7 +20,7 @@ import java.util.Date;
 @NoArgsConstructor
 public class VideoOperateDTO {
     Integer id;
-    Integer uid;
+    Long uid;
     @NotNull(message = "aid不能为空")
     Integer aid;
     @NotNull(message = "pid不能为空")
